requirements.txt
RPLidar
numpy
scipy
pygame
smbus2
pykalman
opencv-python
matplotlib
scikit-learn